requirements.txt
asarPy==1.0.1
mysql-connector-python==9.5.0
python-dotenv==1.2.1
